refactor: replace ActionList with SendLinkActionList for claim flow and remove unused ActionList component
